reorganization with existing DeleteDuplicates, Complement (work in progress)
